+++ crates/lib/src/stdlib/tags/include_tag.rs
@@ -1,13 +1,12 @@
 use std::io::Write;
 
-use liquid_core::compiler::TryMatchToken;
 use liquid_core::error::ResultLiquidExt;
 use liquid_core::Expression;
 use liquid_core::Language;
 use liquid_core::Renderable;
-use liquid_core::Result;
 use liquid_core::Runtime;
 use liquid_core::ValueView;
+use liquid_core::{Error, Result};
 use liquid_core::{ParseTag, TagReflection, TagTokenIter};
 
 #[derive(Debug)]
@@ -17,7 +16,13 @@ struct Include {
 
 impl Renderable for Include {
     fn render_to(&self, writer: &mut dyn Write, runtime: &mut Runtime<'_>) -> Result<()> {
-        let name = self.partial.evaluate(runtime)?.render().to_string();
+        let value = self.partial.evaluate(runtime)?;
+        if !value.is_scalar() {
+            return Error::with_msg("Can only `include` strings")
+                .context("partial", format!("{}", value.source()))
+                .into_err();
+        }
+        let name = value.to_kstr().into_owned();
         runtime.run_in_named_scope(name.clone(), |mut scope| -> Result<()> {
             let partial = scope
                 .partials()
@@ -59,20 +64,12 @@ impl ParseTag for IncludeTag {
         mut arguments: TagTokenIter<'_>,
         _options: &Language,
     ) -> Result<Box<dyn Renderable>> {
-        let name = arguments.expect_next("Identifier or literal expected.")?;
+        let partial = arguments.expect_next("Identifier or literal expected.")?;
 
-        // This may accept strange inputs such as `{% include 0 %}` or `{% include filterchain | filter:0 %}`.
-        // Those inputs would fail anyway by there being not a path with those names so they are not a big concern.
-        let name = match name.expect_literal() {
-            // Using `to_kstr()` on literals ensures `Strings` will have their quotes trimmed.
-            TryMatchToken::Matches(name) => name.to_kstr().to_string(),
-            TryMatchToken::Fails(name) => name.as_str().to_string(),
-        };
+        let partial = partial.expect_value().into_result()?;
 
         // no more arguments should be supplied, trying to supply them is an error
         arguments.expect_nothing()?;
-
-        let partial = Expression::with_literal(name);
 
         Ok(Box::new(Include { partial }))
     }
@@ -157,31 +154,6 @@ mod test {
     #[test]
     fn include_tag_quotes() {
         let text = "{% include 'example.txt' %}";
-        let mut options = options();
-        options
-            .filters
-            .register("size".to_string(), Box::new(SizeFilterParser));
-        let template = compiler::parse(text, &options)
-            .map(interpreter::Template::new)
-            .unwrap();
-
-        let partials = partials::OnDemandCompiler::<TestSource>::empty()
-            .compile(::std::sync::Arc::new(options))
-            .unwrap();
-        let mut runtime = RuntimeBuilder::new()
-            .set_partials(partials.as_ref())
-            .build();
-        runtime.stack_mut().set_global("num", Value::scalar(5f64));
-        runtime
-            .stack_mut()
-            .set_global("numTwo", Value::scalar(10f64));
-        let output = template.render(&mut runtime).unwrap();
-        assert_eq!(output, "5 wat wot");
-    }
-
-    #[test]
-    fn include_non_string() {
-        let text = "{% include example.txt %}";
         let mut options = options();
         options
             .filters